requirements.txt
absl-py==0.10.0
astor==0.8.1
certifi==2020.6.20
chardet==3.0.4
gast==0.2.2
google-pasta==0.2.0
gpt-2-simple==0.7.1
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.19.1
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
regex==2020.7.14
requests==2.24.0
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
toposort==1.5
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
